+++ airflow/models/__init__.py
@@ -20,7 +20,6 @@ from airflow.models.base import ID_LEN, Base
 from airflow.models.baseoperator import BaseOperator, BaseOperatorLink
 from airflow.models.connection import Connection
 from airflow.models.dag import DAG, DagModel, DagTag
-from airflow.models.dagbag import DagBag
 from airflow.models.dagpickle import DagPickle
 from airflow.models.dagrun import DagRun
 from airflow.models.errors import ImportError  # pylint: disable=redefined-builtin
