+++ airflow/serialization/serialized_objects.py
@@ -35,7 +35,7 @@ from airflow.serialization.helpers import serialize_template_field
 from airflow.serialization.json_schema import Validator, load_dag_schema
 from airflow.settings import json
 from airflow.utils.module_loading import import_string
-from airflow.www.utils import get_python_source
+from airflow.utils.source_utils import get_python_source
 
 log = logging.getLogger(__name__)
 FAILED = 'serialization_failed'
