+++ scripts/perf/scheduler_ops_metrics.py
@@ -24,7 +24,8 @@ import pandas as pd
 from airflow import settings
 from airflow.configuration import conf
 from airflow.jobs.scheduler_job import SchedulerJob
-from airflow.models import DagBag, DagModel, DagRun, TaskInstance
+from airflow.models import DagModel, DagRun, TaskInstance
+from airflow.models.dagbag import DagBag
 from airflow.utils import timezone
 from airflow.utils.state import State
