+++ tests/api/common/experimental/test_delete_dag.py
@@ -21,7 +21,9 @@ import unittest
 from airflow import models
 from airflow.api.common.experimental.delete_dag import delete_dag
 from airflow.exceptions import DagNotFound
+from airflow.models.dagbag import DagBag
 from airflow.operators.dummy_operator import DummyOperator
+from airflow.utils import dag_cleaner
 from airflow.utils.dates import days_ago
 from airflow.utils.session import create_session
 from airflow.utils.state import State
@@ -38,12 +40,12 @@ IE = models.ImportError
 class TestDeleteDAGCatchError(unittest.TestCase):
 
     def setUp(self):
-        self.dagbag = models.DagBag(include_examples=True)
+        self.dagbag = DagBag(include_examples=True)
         self.dag_id = 'example_bash_operator'
         self.dag = self.dagbag.dags[self.dag_id]
 
     def tearDown(self):
-        self.dag.clear()
+        dag_cleaner.clear(self.dag)
 
     def test_delete_dag_non_existent_dag(self):
         with self.assertRaises(DagNotFound):
