+++ tests/cli/commands/test_variable_command.py
@@ -19,16 +19,16 @@
 import os
 import unittest
 
-from airflow import models
 from airflow.cli import cli_parser
 from airflow.cli.commands import variable_command
 from airflow.models import Variable
+from airflow.models.dagbag import DagBag
 
 
 class TestCliVariables(unittest.TestCase):
     @classmethod
     def setUpClass(cls):
-        cls.dagbag = models.DagBag(include_examples=True)
+        cls.dagbag = DagBag(include_examples=True)
         cls.parser = cli_parser.get_parser()
 
     def test_variables(self):
