+++ tests/models/test_serialized_dag.py
@@ -21,7 +21,7 @@
 import unittest
 
 from airflow import example_dags as example_dags_module
-from airflow.models import DagBag
+from airflow.models.dagbag import DagBag
 from airflow.models.dagcode import DagCode
 from airflow.models.serialized_dag import SerializedDagModel as SDM
 from airflow.serialization.serialized_objects import SerializedDAG
