+++ tests/sensors/test_external_task_sensor.py
@@ -22,8 +22,9 @@ import pytest
 
 from airflow import exceptions, settings
 from airflow.exceptions import AirflowException, AirflowSensorTimeout
-from airflow.models import DagBag, TaskInstance
+from airflow.models import TaskInstance
 from airflow.models.dag import DAG
+from airflow.models.dagbag import DagBag
 from airflow.operators.bash import BashOperator
 from airflow.operators.dummy_operator import DummyOperator
 from airflow.sensors.external_task_sensor import ExternalTaskMarker, ExternalTaskSensor
