+++ tests/sensors/test_weekday_sensor.py
@@ -23,8 +23,9 @@ from parameterized import parameterized
 
 from airflow.contrib.utils.weekday import WeekDay
 from airflow.exceptions import AirflowSensorTimeout
-from airflow.models import DagBag, TaskFail, TaskInstance
+from airflow.models import TaskFail, TaskInstance
 from airflow.models.dag import DAG
+from airflow.models.dagbag import DagBag
 from airflow.sensors.weekday_sensor import DayOfWeekSensor
 from airflow.settings import Session
 from airflow.utils.timezone import datetime
