+++ tests/serialization/test_dag_serialization.py
@@ -29,8 +29,9 @@ from dateutil.relativedelta import FR, relativedelta
 from parameterized import parameterized
 
 from airflow.hooks.base_hook import BaseHook
-from airflow.models import DAG, Connection, DagBag, TaskInstance
+from airflow.models import DAG, Connection, TaskInstance
 from airflow.models.baseoperator import BaseOperator
+from airflow.models.dagbag import DagBag
 from airflow.operators.bash import BashOperator
 from airflow.operators.subdag_operator import SubDagOperator
 from airflow.serialization.json_schema import load_dag_schema_dict
