+++ tests/www/api/experimental/test_dag_runs_endpoint.py
@@ -21,7 +21,8 @@ import unittest
 from parameterized import parameterized_class
 
 from airflow.api.common.experimental.trigger_dag import trigger_dag
-from airflow.models import DagBag, DagRun
+from airflow.models import DagRun
+from airflow.models.dagbag import DagBag
 from airflow.models.serialized_dag import SerializedDagModel
 from airflow.settings import Session
 from airflow.www import app as application
